shotlist options export tab mockup
