+++ backend/src/main/resources/application.properties
@@ -3,7 +3,7 @@ quarkus.http.cors.origins=*
 quarkus.http.cors.methods=GET, POST, PUT, DELETE
 quarkus.http.cors.headers=Content-Type, Authorization
 
-quarkus.hibernate-orm.database.generation=update
+quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=false
 quarkus.hibernate-orm.sql-load-script=import.sql
