+++ frontend/src/components/dialog/shotlistOptionsDialog/exportTab/exportTab.scss
@@ -1,0 +1,49 @@
+div.shotlistOptionsDialogExportTab{
+  h2{
+    font-weight: 600;
+    margin-block: .5rem;
+  }
+
+  .filters{
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+    .filter{
+      display: flex;
+      justify-content: space-between;
+      gap: 1rem;
+
+      .left{
+        display: flex;
+        gap: .5rem;
+      }
+    }
+
+    button.addFilter{
+      border: .1rem solid var(--default-interactable-border);
+      width: fit-content;
+      color: var(--text-30);
+      padding-block: .3rem;
+      font-size: .9rem;
+
+      &:hover{
+        border-color: var(--default-interactable-hover-border);
+        color: var(--text);
+      }
+    }
+  }
+
+  button.export{
+    padding: .4rem .8rem;
+    color: var(--contrast);
+    background-color: var(--accent);
+    font-weight: bold;
+    margin-top: 2rem;
+    font-size: .95rem;
+
+    &:hover{
+      background-color: var(--accent-55);
+    }
+  }
+}
