+++ frontend/src/components/dialog/shotlistOptionsDialog/shotlistOptionsDialoge.tsx
@@ -4,7 +4,7 @@ import * as Dialog from '@radix-ui/react-dialog';
 import React, {useEffect, useState} from 'react';
 import "./shotlistOptionsDialog.scss"
 import {Popover, Separator, Tabs, VisuallyHidden} from "radix-ui"
-import {ChevronDown, FileDown, List, Plus, Type, Users, X} from "lucide-react"
+import {ChevronDown, File, FileDown, List, Plus, Type, Users, X, ListOrdered} from "lucide-react"
 import ShotAttributeDefinition from "@/components/shotAttributeDefinition/shotAttributeDefinition"
 import {
     AnySceneAttributeDefinition,
@@ -23,10 +23,11 @@ import {
 } from "../../../../lib/graphql/generated"
 import Image from "next/image"
 import SceneAttributeDefinition from "@/components/sceneAttributeDefinition/sceneAttributeDefinition"
-import ReactPDF, {pdf, PDFDownloadLink} from '@react-pdf/renderer';
+import ReactPDF, {Page, pdf, PDFDownloadLink} from '@react-pdf/renderer';
 import PDFExport from "@/components/PDFExport"
 import {wuTime} from "@yanikkendler/web-utils/dist"
 import {useRouter} from "next/navigation"
+import ExportTab from "@/components/dialog/shotlistOptionsDialog/exportTab/exportTab"
 
 export type ShotlistOptionsDialogPage = "attributes" | "collaborators" | "export"
 
@@ -273,77 +274,6 @@ export default function ShotlistOptionsDialog({isOpen, setIsOpen, selectedPage, 
         }
     }
 
-    async function exportPDF() {
-            const {data, error, loading} = await client.query({
-                    query: gql`
-                        query shotlistForExport($id: String!) {
-                            shotlist(id: $id){
-                                id
-                                name
-                                scenes{
-                                    id
-                                    position
-                                    attributes{
-                                        id
-                                        definition{id, name, position}
-
-                                        ... on SceneSingleSelectAttributeDTO{
-                                            singleSelectValue{id,name}
-                                        }
-
-                                        ... on SceneMultiSelectAttributeDTO{
-                                            multiSelectValue{id,name}
-                                        }
-                                        ... on SceneTextAttributeDTO{
-                                            textValue
-                                        }
-                                    }
-                                    shots {
-                                        id
-                                        position
-                                        attributes{
-                                            id
-                                            definition{id, name, position}
-
-                                            ... on ShotSingleSelectAttributeDTO{
-                                                singleSelectValue{id,name}
-                                            }
-
-                                            ... on ShotMultiSelectAttributeDTO{
-                                                multiSelectValue{id,name}
-                                            }
-                                            ... on ShotTextAttributeDTO{
-                                                textValue
-                                            }
-                                        }
-                                    }
-                                }
-                                sceneAttributeDefinitions{
-                                    id
-                                    name
-                                }
-                                shotAttributeDefinitions{
-                                    id
-                                    name
-                                }
-                            }
-                        }`,
-                    variables: {id: shotlistId}
-                }
-            )
-
-        console.log(data)
-
-        const blob = await pdf(<PDFExport data={data.shotlist}/>).toBlob()
-
-        const url = URL.createObjectURL(blob)
-        const link = document.createElement('a')
-        link.href = url
-        link.download = `shotly-${shotlist.name}-${wuTime.toFullDateTimeString(Date.now())}.pdf`
-        link.click()
-        URL.revokeObjectURL(url)
-    }
-
     function runRefreshShotlistCheck(){
         let currentAttributeData = JSON.stringify(shotAttributeDefinitions) + JSON.stringify(sceneAttributeDefinitions)
 
@@ -490,11 +420,10 @@ export default function ShotlistOptionsDialog({isOpen, setIsOpen, selectedPage, 
                                 </Tabs.Root>
                             </Tabs.Content>
                             <Tabs.Content value={"collaborators"} className={"content"}>
-                                collaborators
+                                This feature is under development and will be available soon.
                             </Tabs.Content>
                             <Tabs.Content value={"export"} className={"content"}>
-
-                                <button onClick={exportPDF}>test export</button>
+                                <ExportTab shotlist={shotlist}/>
                             </Tabs.Content>
                         </Tabs.Root>
                     </div>
